Added valiator crate. Refactored routes. Initial validation file commit.
