+++ blackrose-website-backend/Cargo.toml
@@ -22,3 +22,4 @@ serde = { version = "1.0.160", features = ["derive"] }
 serde_json = "1.0.96"
 thiserror = "1.0.40"
 tokio = { version = "1.27.0", features = ["tokio-macros", "macros", "rt-multi-thread"] }
+validator = { version = "0.16.0", features = ["derive"] }
