+++ blackrose-website-backend/src/authentication.rs
@@ -1,0 +1,4 @@
+pub mod auth;
+pub mod login;
+pub mod validation;
+pub mod registration;
